full_stack/requirements.txt
Django==4.0.4
pandas==1.4.2
plotly==5.7.0
psycopg2-binary==2.9.3
django-cors-headers==3.11.0
PyJWT==2.3.0
djangorestframework-simplejwt==5.0.0